+++ src/pca.rs
@@ -1,16 +1,16 @@
 use crate::utils;
 use na::DMatrix;
 
-fn compute_covariance_matrix(x: &DMatrix<f32>, unbiased: bool) -> DMatrix<f32> {
+fn compute_covariance_matrix(x: &DMatrix<f64>, unbiased: bool) -> DMatrix<f64> {
     let bias = if unbiased { 1. } else { 0. };
-    let nrows = x.nrows() as f32;
+    let nrows = x.nrows() as f64;
     let means = x.row_mean();
     let centered = utils::subtract_row_from_matrix(&x, &means);
     let covariance_matrix = (&centered.transpose() * &centered) / (nrows - bias);
     covariance_matrix
 }
 
-fn eigen_sign_correction(mut u: DMatrix<f32>) -> DMatrix<f32> {
+fn eigen_sign_correction(mut u: DMatrix<f64>) -> DMatrix<f64> {
     let nrows = u.nrows();
     let mut sign;
     for i in 0..nrows {
@@ -25,7 +25,7 @@ fn eigen_sign_correction(mut u: DMatrix<f32>) -> DMatrix<f32> {
     u
 }
 
-fn compute_eigenvectors(matrix: DMatrix<f32>) -> DMatrix<f32> {
+fn compute_eigenvectors(matrix: DMatrix<f64>) -> DMatrix<f64> {
     let eigen = matrix.symmetric_eigen();
     let mut indices: Vec<usize> = (0..eigen.eigenvalues.len()).collect();
     indices.sort_by(|&i, &j| {
@@ -43,7 +43,7 @@ fn compute_eigenvectors(matrix: DMatrix<f32>) -> DMatrix<f32> {
     u_corrected
 }
 
-pub fn compute_pca(matrix: &DMatrix<f32>) -> DMatrix<f32> {
+pub fn compute_pca(matrix: &DMatrix<f64>) -> DMatrix<f64> {
     let covariance_matrix = compute_covariance_matrix(matrix, false);
     let pca_matrix = compute_eigenvectors(covariance_matrix);
     pca_matrix
