+++ src/preprocessing.rs
@@ -2,9 +2,9 @@ use ordered_float::OrderedFloat;
 use std::collections::BTreeMap;
 
 #[derive(PartialEq, Eq, PartialOrd, Ord)]
-struct OrderedPoint(OrderedFloat<f32>, OrderedFloat<f32>, OrderedFloat<f32>);
+struct OrderedPoint(OrderedFloat<f64>, OrderedFloat<f64>, OrderedFloat<f64>);
 
-fn to_ordered_point(num: [f32; 3]) -> OrderedPoint {
+fn to_ordered_point(num: [f64; 3]) -> OrderedPoint {
     OrderedPoint(
         OrderedFloat(num[0]),
         OrderedFloat(num[1]),
@@ -12,14 +12,14 @@ fn to_ordered_point(num: [f32; 3]) -> OrderedPoint {
     )
 }
 
-fn from_ordered_point(num: OrderedPoint) -> [f32; 3] {
+fn from_ordered_point(num: OrderedPoint) -> [f64; 3] {
     [num.0.into(), num.1.into(), num.2.into()]
 }
 
 fn rgb_to_yuv(rgb: [u8; 3]) -> [u8; 3] {
-    let r = rgb[0] as f32;
-    let g = rgb[1] as f32;
-    let b = rgb[2] as f32;
+    let r = rgb[0] as f64;
+    let g = rgb[1] as f64;
+    let b = rgb[2] as f64;
     let y = (0.2126 * r + 0.2126 * g + 0.0722 * b) + 0.;
     let u = (-0.1146 * r + -0.3854 * g + 0.5000 * b) + 128.;
     let v = (0.5000 * r + -0.4542 * g + -0.0468 * b) + 128.;
@@ -31,7 +31,7 @@ fn rgb_to_yuv(rgb: [u8; 3]) -> [u8; 3] {
 }
 
 fn mean_colors(colors: Vec<[u8; 3]>) -> [u8; 3] {
-    let vec_len = colors.len() as f32;
+    let vec_len = colors.len() as f64;
     let mut vec_sum = (0, 0, 0);
     for color in colors {
         vec_sum.0 += color[0];
@@ -39,9 +39,9 @@ fn mean_colors(colors: Vec<[u8; 3]>) -> [u8; 3] {
         vec_sum.2 += color[2];
     }
     let mean = (
-        vec_sum.0 as f32 / vec_len,
-        vec_sum.1 as f32 / vec_len,
-        vec_sum.2 as f32 / vec_len,
+        vec_sum.0 as f64 / vec_len,
+        vec_sum.1 as f64 / vec_len,
+        vec_sum.2 as f64 / vec_len,
     );
     let rounded_mean = [
         mean.0.round() as u8,
@@ -52,9 +52,9 @@ fn mean_colors(colors: Vec<[u8; 3]>) -> [u8; 3] {
 }
 
 pub fn preprocess_point_cloud(
-    points: Vec<[f32; 3]>,
+    points: Vec<[f64; 3]>,
     colors: Vec<[u8; 3]>,
-) -> (Vec<[f32; 3]>, Vec<[u8; 3]>) {
+) -> (Vec<[f64; 3]>, Vec<[u8; 3]>) {
     let mut points_map: BTreeMap<OrderedPoint, Vec<[u8; 3]>> = BTreeMap::new();
     for i in 0..points.len() {
         let point = to_ordered_point(points[i]);
